+++ test/action.test.js
@@ -1,7 +1,9 @@
 /* eslint-env mocha */
 const assert = require('chai').assert
 const fs = require('fs')
-const { spawnSync } = require('child_process')
+const runAction = require('./helpers/run-action.js')
+const initData = require('./helpers/init-data.js')
+const steps = ['pre', 'main']
 const deepEqual = require('fast-deep-equal')
 if (!process.env.GITHUB_ACTIONS) {
   const path2Env = `${process.cwd()}/test/.env`
@@ -18,7 +20,7 @@ describe('action-collmex-scraper', function () {
   describe('at first run', function () {
     before(async function () {
       this.timeout(30000)
-      runAction()
+      runAction(steps)
     })
     it('should fetch data from Collmex docs', function () {
       assert.equal(fs.existsSync(filePath), true)
@@ -38,7 +40,7 @@ describe('action-collmex-scraper', function () {
     before(async function () {
       oldData = require(filePath)
       this.timeout(30000)
-      runAction()
+      runAction(steps)
     })
     it('should not update data', function () {
       const newData = require(filePath)
@@ -53,7 +55,7 @@ describe('action-collmex-scraper', function () {
       }
       fs.writeFileSync(filePath, JSON.stringify(oldData, null, 2), 'utf-8')
       this.timeout(30000)
-      runAction()
+      runAction(steps)
     })
     it('should create a backup file', function () {
       const backupPath = `${filePath.slice(0, -5)}.backup${filePath.slice(-5)}`
@@ -70,33 +72,3 @@ describe('action-collmex-scraper', function () {
     })
   })
 })
-
-function initData (filePath) {
-  return () => {
-    const dirPath = filePath.split('/').slice(0, filePath.split('/').length - 1).join('/')
-    if (fs.existsSync(dirPath)) {
-      cleanData(dirPath)
-    } else {
-      fs.mkdirSync(dirPath)
-    }
-  }
-}
-
-function cleanData (dirPath) {
-  const filePaths = fs.readdirSync(dirPath, { withFileTypes: true }).filter(dirent => !dirent.isDirectory()).map(dirent => `${dirPath}/${dirent.name}`)
-  for (const filePath of filePaths) {
-    fs.unlinkSync(filePath)
-  }
-}
-
-function runAction () {
-  try {
-    console.log('INFO: running action...')
-    spawnSync('node', ['src/pre'], { stdio: 'inherit' })
-    spawnSync('node', ['src/main'], { stdio: 'inherit' })
-  } catch (err) {
-    console.log('ERROR: an error occured...')
-    console.log(err)
-    process.exit()
-  }
-}
